Move purger to root level package
